Add more precompiled assets
The last commit ended up only including a subset of the precompiled
assets. To truly get them all we need to compile using:

  RAILS_ENV=production bundle exec rake assets:precompile

Note that some tweaks were required to the application config
(explicitly listing the stylesheet files we wanted to precompile, and
renaming them to follow the expected conventions).